logic_manager changes sprite of sample + other small things
the logic manager now has reference to the sample. Upon pressing t ("Test") the logic manager calls the function "bacteria_transferred" in the sample. there the sprite is changed.

also the petri dish sprite was changed to pipetting since this is the initial and most important minigame we will work on.