scene_logic implemented
player, sample and logic_manager are not destroyed on load but are deactivated when a new scene is opened. when returning to main (scene) the objects are reactivated and their informaiton is preserved. To end the minigame, some script inside the minigame just has to call the public function return_to_lab with the Success = (T/F) argument.