adding centrifugation icon and major bug fixes
removed the prefab as I realized they were not the right tool for that.

removed the object2scene scripts and merged the two responsible scripts into one (at the cost of much more redundancy in the game but okay).

had to play around with the instructor because there was a weird error. working again now.